moved more func to module
